Refactor server tests in TypeScript
